audio_sentiment_analysis/model/requirements.txt
huggingface
openai-whisper
setuptools-rust
pyannote.audio
python-dotenv
pydub